cleanup started
git-svn-id: https://svn.code.sf.net/p/ultravnc/code/UltraVNC%20Project%20Root/UltraVNC@1185 0bd26501-c230-0410-ab9e-a7946180fe3b

+++ ZipUnZip32/ZipUnzip32.cpp
@@ -146,7 +146,7 @@ bool CZipUnZip32::ZipDirectory(char* szRootDir, char* szDirectoryToZip, char* sz
 	sz = szFileList + cc;
 	for (i = 0; i < m_ZpZCL.argc; i++)
 	{
-		cc = strlen(szDirectoryToZip);
+		cc = (int)strlen(szDirectoryToZip);
 		strcpy(sz, szDirectoryToZip);
 		index[i] = sz;
 		sz += (cc + 1);
